user_service/requirements.txt
Django==5.1.6
djangorestframework==3.16.0
python-dotenv==1.0.1
requests==2.32.3
psycopg2==2.9.10
Pillow==11.1.0